moved selectors back to preprocessing to avoid name clash with stdlib module
